+++ src/constants/og.ts
@@ -1,0 +1,53 @@
+export const OpenGraphImages = {
+  showcase: {
+    url: 'og/default.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description: 'Top iOS apps using Expo Open Source Software in production',
+  },
+  brand: {
+    url: 'og/default.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description: 'Brand assets',
+  },
+  game: {
+    url: 'og/default.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description: 'Games',
+  },
+  lego: {
+    url: 'og/lego.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description:
+      'Modest Evan Bacon standing humbly amongst his Marvel Lego creations at Salt Lake City Comic Con',
+  },
+  talks: {
+    url: 'og/talks.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description:
+      'Sleepy Evan Bacon sleeping before giving his App.js Conference Talk in 2019',
+  },
+  code: {
+    url: 'og/code.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description: 'Focused Evan Bacon writing code with an intense stare',
+  },
+  media: {
+    url: 'og/media.jpg',
+    type: 'image/jpeg',
+    width: 1280,
+    height: 720,
+    description: 'Happy Evan Bacon interviewed at Comic Con',
+  },
+};
