+++ tsconfig.json
@@ -1,19 +1,11 @@
 {
   "compilerOptions": {
-    "typeRoots": [
-      "./global.d.ts"
-    ],
+    "typeRoots": ["./global.d.ts"],
+    "baseUrl": ".",
     "paths": {
-      "@/*": [
-        "./src/*"
-      ]
+      "@/*": ["./src/*"]
     }
   },
   "extends": "expo/tsconfig.base",
-  "include": [
-    "**/*.ts",
-    "**/*.tsx",
-    ".expo/types/**/*.ts",
-    "expo-env.d.ts"
-  ]
+  "include": ["**/*.ts", "**/*.tsx", ".expo/types/**/*.ts", "expo-env.d.ts"]
 }
